% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{eca}
\alias{eca}
\alias{eca-package}
\title{The eca package: A package for estimating catch at age.}
\description{
The eca package provides ..
}
\section{Introduction}{

In order to set quotas for the catch of commercial fish species such as cod, complex population dynamics models have been developed. 
An essential input to these models is the age distribution of the fish caught every year. This is estimated from fish sampled from 
boats as they land their catch. The major problem with this data is that although the number of fish sampled is large, there is a very 
large within-boat correlation, and the number of boats sampled is relatively small. Also the age distribution changes in space and time, 
and with other covariates as the type of gear used by the boats, season etc.

The Norwegian Computing Center and the Institue of Marine Research have over years developed a Bayesian hierarchical model to estimate
the catch at age. The model allows us to estimate the age distribution in areas where there are few or no samples, and it also give 
a measure of the uncertainty in the estimation.

The model is implemented in C with an R interface.
}

\section{The model}{

Briefly describe the model.
}

\section{References}{

\href{http://www.nrcresearchpress.com/doi/abs/10.1139/cjfas-2012-0075#.WjE01LBrwuW}{Hirst et al 2012}
}

