% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.list.functions.R
\name{make.hszList}
\alias{make.hszList}
\title{Make input list for haulsize data}
\usage{
make.hszList(ageList)
}
\arguments{
\item{ageList}{list with age data}
}
\description{
Make input list for haulsize data
}
