% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.list.functions.R
\name{make.length.intervals}
\alias{make.length.intervals}
\title{Create length interval information}
\usage{
make.length.intervals(data, resolution)
}
\description{
Create length interval information
}
