% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.list.functions.R
\name{make.lgaList}
\alias{make.lgaList}
\title{Make input list for lga data}
\usage{
make.lgaList(data, common)
}
\arguments{
\item{data}{list with input data}

\item{common}{list with common parameters}
}
\description{
Make input list for lga data
}
