% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.list.functions.R
\name{make.wglList}
\alias{make.wglList}
\title{Make input list for wgl data}
\usage{
make.wglList(data, common)
}
\arguments{
\item{data}{list with input data}

\item{common}{list with common parameters}
}
\description{
Make input list for wgl data
}
