% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bin.R
\name{read.fit.bin}
\alias{read.fit.bin}
\title{Read binary file from estimation with mcmc samples}
\usage{
read.fit.bin(filename.mcmc1, filename.mcmc2, filename.hsz, stoxdata)
}
\arguments{
\item{stoxdata}{list with input data}

\item{filename}{filename (with correct path) of binary file to read mcmc results from estimation}
}
\description{
Read binary file from estimation with mcmc samples
}
