% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bin.R
\name{read.predict.bin}
\alias{read.predict.bin}
\title{Read binary file from prediction with mcmc samples}
\usage{
read.predict.bin(filename.predict)
}
\arguments{
\item{filename}{filename (with correct path) of binary file to read mcmc results from estimation}
}
\description{
Read binary file from prediction with mcmc samples
}
