% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.fit.R
\name{run.fit}
\alias{run.fit}
\title{Estimate catch-at-age model}
\usage{
run.fit(stoxdata, common, win)
}
\arguments{
\item{stoxdata}{list with input data}

\item{common}{list with common parameters}

\item{win}{indicator variable, =1 if windows platform, 0 otherwise}
}
\description{
Estimate catch-at-age model
}
\keyword{c-program}
\keyword{mcmc}
