% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.pred.R
\name{run.pred}
\alias{run.pred}
\title{Predict from catch-at-age model}
\usage{
run.pred(stoxdata, common, win)
}
\arguments{
\item{stoxdata}{list with input data}

\item{win}{indicator variable, =1 if windows platform, 0 otherwise}
}
\description{
Predict from catch-at-age model
}
\keyword{c-program}
\keyword{mcmc}
